Cleanup KafkaConnectSources
