+++ extensions/kafka-connect/src/main/java/com/hazelcast/jet/kafka/connect/KafkaConnectSources.java
@@ -17,11 +17,7 @@
 package com.hazelcast.jet.kafka.connect;
 
 import com.hazelcast.internal.util.Preconditions;
-import com.hazelcast.jet.core.Processor;
 import com.hazelcast.jet.core.ProcessorMetaSupplier;
-import com.hazelcast.jet.core.ProcessorSupplier;
-import com.hazelcast.jet.kafka.connect.impl.ConnectorWrapper;
-import com.hazelcast.jet.kafka.connect.impl.ReadKafkaConnectP;
 import com.hazelcast.jet.pipeline.Sources;
 import com.hazelcast.jet.pipeline.StreamSource;
 import com.hazelcast.jet.pipeline.StreamStage;
@@ -29,14 +25,11 @@ import com.hazelcast.spi.annotation.Beta;
 import org.apache.kafka.connect.source.SourceRecord;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.net.URL;
-import java.util.Collection;
 import java.util.Properties;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import static com.hazelcast.internal.util.Preconditions.checkRequiredProperty;
+import static com.hazelcast.jet.kafka.connect.impl.ReadKafkaConnectP.processSupplier;
 
 /**
  * Contains factory methods to create a Kafka Connect source.
@@ -87,33 +80,7 @@ public final class KafkaConnectSources {
         }
 
         return Sources.streamFromProcessorWithWatermarks(name, true,
-                eventTimePolicy -> ProcessorMetaSupplier.randomMember(
-                        new ProcessorSupplier() {
-                            private transient ConnectorWrapper connectorWrapper;
-
-                            @Override
-                            public void init(@Nonnull Context context) {
-                                properties.put("tasks.max", Integer.toString(context.localParallelism()));
-                                this.connectorWrapper = new ConnectorWrapper(properties);
-                            }
-
-                            @Override
-                            public void close(@Nullable Throwable error) {
-                                if (connectorWrapper != null) {
-                                    connectorWrapper.stop();
-                                }
-                            }
-
-                            @Nonnull
-                            @Override
-                            public Collection<? extends Processor> get(int count) {
-                                return IntStream.range(0, count)
-                                        .mapToObj(i -> new ReadKafkaConnectP(connectorWrapper, eventTimePolicy))
-                                        .collect(Collectors.toList());
-                            }
-
-                        })
-        );
+                eventTimePolicy -> ProcessorMetaSupplier.randomMember(processSupplier(properties, eventTimePolicy)));
     }
 
 }
